continue

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: bbotelho <bbotelho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/09 13:29:49 by bbotelho          #+#    #+#             */
-/*   Updated: 2023/09/14 11:39:07 by bbotelho         ###   ########.fr       */
+/*   Updated: 2023/10/01 05:38:02 by bbotelho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
